% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shifts vector values to right or left}
\usage{
shift(x, n, invert = FALSE, default = NA)
}
\arguments{
\item{x}{Vector for which to shift values}

\item{n}{Number of places to be shifted.
Positive numbers will shift to the right by default.
Negative numbers will shift to the left by default.
The direction can be inverted by the invert parameter.}

\item{invert}{Whether or not the default shift directions should be inverted.}

\item{default}{The value that should be inserted by default.}
}
\value{
a vector.
}
\description{
Shifts vector values to right or left
}
\examples{
# Lag
shift(c(2,3,5,6,7), n = 1, default = 0)
# Lead
shift(c(2,3,5,6,7), n = -1, default = 0)

}
